refactor(test): extract multimodal framework to tests/utils/multimodal.py
Move MultimodalModelProfile, TopologyConfig, and make_multimodal_configs
to a dedicated utility module. Evolve the framework to support
per-topology marks, timeout, and VRAM overrides.

Migrate 3 hand-written configs to the framework:
- multimodal_e_pd_qwen → mm_e_pd_qwen3-vl-2b (pre_merge)
- multimodal_disagg_qwen → mm_epd_qwen3-vl-2b (pre_merge)
- multimodal_agg_qwen → mm_agg_qwen2.5-vl-7b (post_merge)

Non-migratable configs (video, audio, tool calling, frontend-decoding,
llava, base64) remain hand-written.
